# PERF: Speedup validate_dtype