--- a/tests/test_dtypes.py
+++ b/tests/test_dtypes.py
@@ -5,6 +5,7 @@ from .conftest import gdal_version
 
 import rasterio
 from rasterio import (
+    bool_,
     ubyte,
     uint8,
     uint16,
@@ -82,6 +83,12 @@ def test_get_minimum_dtype():
     assert get_minimum_dtype(np.array([-1, 0, 128], dtype=int)) == int16
     assert get_minimum_dtype(np.array([-1, 0, 100000], dtype=int)) == int32
     assert get_minimum_dtype(np.array([-1.5, 0, 1.5], dtype=np.float64)) == float32
+
+    assert get_minimum_dtype([-1.5, np.nan, np.inf, -np.inf]) == float32
+    assert get_minimum_dtype(np.array([-1.5, np.nan, np.inf, -np.inf], dtype=np.float64)) == float32
+    assert get_minimum_dtype([-np.inf, np.inf]) == float32
+    assert get_minimum_dtype([]) == bool_
+    assert get_minimum_dtype(np.array([], dtype=int)) == bool_
 
 
 def test_get_minimum_dtype__int64():