Continuing from #2716, we can also avoid creating an array in validate_dtype.

```python
# Fast path (numpy array is exactly the dtype we want) is roughly the same
In [49]: %timeit validate_dtype(nvals, ['int32'])
1.82 µs ± 6.06 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

In [50]: %timeit validate_dtype2(nvals, ['int32'])
1.88 µs ± 33.3 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

# Array, but dtype is not valid (needing to call get_minimum_dtype) basically the same.
In [66]: %timeit validate_dtype(nvals, ['float32'])
4.77 µs ± 366 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

In [67]: %timeit validate_dtype2(nvals, ['float32'])
4.46 µs ± 56 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

# Slow path (non-array paths) are much faster
In [64]: %timeit validate_dtype(vals, ['int32'])
14.1 µs ± 43.8 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

In [65]: %timeit validate_dtype2(vals, ['int32'])
6.37 µs ± 44.2 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

Question 1: There is a comment at the top of the module that talks about not importing numpy. Is that still relevant?
https://github.com/rasterio/rasterio/blob/main/rasterio/dtypes.py#L3

Question 2: Would it be desired for dtype objects or strings to be accepted as types? 
for example: `validate_dtype(arr, [np.uint8, 'int16'])`