@groutr 2023-01-20
And Question 3: Should this function return false if an array is passed and the dtype is not valid (without calling `get_minimum_dtype`)?
ie
```python
    if is_ndarray(values):
        return values.dtype.name in valid_dtypes
    else:
        return get_minimum_dtype(values) in valid_dtypes
```