@snowman2 2023-01-21
Thoughts about:
```python
if is_ndarray(values) and values.dtype.name in valid_dtypes:
    return True
return get_minimum_dtype(values) in valid_dtypes
```