@snowman2 2023-01-21
> Question 1: There is a comment at the top of the module that talks about not importing numpy. Is that still relevant?

It doesn't appear to be