@snowman2 2023-01-21
> And Question 3: Should this function return false if an array is passed and the dtype is not in valid_dtypes (without calling get_minimum_dtype)?

I believe that the code in it's current state validates that the input array can be safety written to the desired output dtype that the user selected. So, the behavior is as expected.