@groutr 2023-01-21
> Thoughts about:
> 
> ```python
> if is_ndarray(values) and values.dtype.name in valid_dtypes:
>     return True
> return get_minimum_dtype(values) in valid_dtypes
> ```

I have no opinion here. The rest of the module uses the if-else pattern. Happy to change to whatever is preferred.

> > Question 1: There is a comment at the top of the module that talks about not importing numpy. Is that still relevant?
> 
> It doesn't appear to be

Alright, I'll remove that line from the docstring.

> > And Question 3: Should this function return false if an array is passed and the dtype is not in valid_dtypes (without calling get_minimum_dtype)?
> 
> I believe that the code in it's current state validates that the input array can be safety written to the desired output dtype that the user selected. So, the behavior is as expected.

I think my counterexample above shows that `validate_dtype` cannot reliably be used in this way.  That seems to be more inline with stated purpose of `can_cast_dtype`.