@groutr 2023-01-21
> > Thoughts about:
> > ```python
> > if is_ndarray(values) and values.dtype.name in valid_dtypes:
> >     return True
> > return get_minimum_dtype(values) in valid_dtypes
> > ```
> 
> This change has a minor tweak that I believe preserves the original behavior while also adding performance enhancements (untested).

I doubt there would be any difference in performance since both forms compile to identical bytecode instructions.
```python
In [106]: def ifelse(x):
     ...:     if x:
     ...:         return False
     ...:     else:
     ...:         return True
     ...:

In [107]: def ifret(x):
     ...:     if x:
     ...:         return False
     ...:     return True
In [104]: dis.dis(ifelse)
  2           0 LOAD_FAST                0 (x)
              2 POP_JUMP_IF_FALSE        4 (to 8)

  3           4 LOAD_CONST               1 (False)
              6 RETURN_VALUE

  5     >>    8 LOAD_CONST               2 (True)
             10 RETURN_VALUE

In [105]: dis.dis(ifret)
  2           0 LOAD_FAST                0 (x)
              2 POP_JUMP_IF_FALSE        4 (to 8)

  3           4 LOAD_CONST               1 (False)
              6 RETURN_VALUE

  4     >>    8 LOAD_CONST               2 (True)
             10 RETURN_VALUE
```