@snowman2 2023-01-21
> I doubt there would be any difference in performance since both forms compile to identical bytecode instructions.

Please look again - the change doesn't have to do with if/else.