@snowman2 2023-01-23
@groutr, it looks like you have stumbled upon multiple potential improvements in this PR. That is great :+1:

To help get the changes reviewed and approved in an efficient manner I recommend breaking the changes out into separate PRs. This also helps to get some changes merged in without needing to wait for other changes that are in discussion.

I suggest moving these changes to separate PRs:
- check finite values range in get_minimum_dtype
- can_cast_dtype improvements
- removing the note at the top

Also, for question 2 & 3, I suggest opening new issues or PRs with potential implementations to further the discussions as those sound like valid topics to dig into.

Thank you for your patience and for your contributions :+1: